switch to turso database

+++ templates/summary_templ.go
@@ -12,7 +12,7 @@ import (
 	"fmt"
 )
 
-func Summary(score int32) templ.Component {
+func Summary(score int64) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
